+++ doc/examples/blog-app/lib/screens/stats_screen.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 import 'package:flutter/material.dart';
-import 'package:smart_cache/smart_cache.dart';
+import 'package:cache_nexus/cache_nexus.dart';
 import '../services/cache_service.dart';
 
 class StatsScreen extends StatefulWidget {